fix: validaciones

+++ src/controllers/agente.js
@@ -6,8 +6,7 @@ const { encrypt } = require("./auth");
 const get = async (req, res) => {
   try {
     const agente = await Usuario.findAll({
-      
-      include: [{ model: Rol,where: { rol: "Trabajador" }, }],
+      include: [{ model: Rol, where: { rol: "Trabajador" } }],
     });
     return res.status(200).json({ data: agente });
   } catch (error) {
@@ -21,6 +20,20 @@ const post = async (req, res) => {
   try {
     const { cod_rol, password, nombre, dni, celular, correo, oficina } =
       req.body;
+
+    // Verificar si el correo electrónico ya está en uso
+    const emailInUse = await Usuario.findOne({ where: { correo: correo } });
+    if (emailInUse) {
+      return res
+        .status(400)
+        .json({ msg: "El correo ya está en uso." });
+    }
+
+    // Verificar si el DNI ya está en uso
+    const dniInUse = await Usuario.findOne({ where: { dni: dni } });
+    if (dniInUse) {
+      return res.status(400).json({ msg: "El DNI ya está en uso." });
+    }
 
     let nuevoUsuario = {
       nombre: nombre,
@@ -44,6 +57,25 @@ const update = async (req, res) => {
   let id = req.params.id;
   try {
     const { cod_rol, password, nombre, dni, celular, correo } = req.body;
+    const usuario = await Usuario.findOne({
+      where: { cod_permiso: id },
+    });
+
+    if (!usuario) {
+      return res.status(404).json({ msg: "No se encontró el usuario." });
+    }
+    const existingEmail = await db.models.Usuario.findOne({ where: { correo } });
+
+    if (existingEmail && existingEmail.cod_usuario !== cod_usuario) {
+      return res.status(400).json({ msg: "El correo ya está en uso." });
+    }
+
+    // Check if the new DNI is already in use.
+    const existingDni = await db.models.Usuario.findOne({ where: { dni } });
+
+    if (existingDni && existingDni.cod_usuario !== cod_usuario) {
+      return res.status(400).json({ msg: "El DNI ya está en uso." });
+    }
 
     let nuevoUsuario = {
       nombre: nombre,
@@ -71,6 +103,13 @@ const delte = async (req, res) => {
   let id = req.params.id;
 
   try {
+    const usuario = await Usuario.findOne({
+      where: { cod_usuario: id },
+    });
+
+    if (!usuario) {
+      return res.status(404).json({ msg: "No se encontró el usuario." });
+    }
     await Usuario.destroy({
       where: { cod_usuario: id },
     });
@@ -80,5 +119,11 @@ const delte = async (req, res) => {
     res.status(500).json({ msg: "No se pudo eliminar al trabajador." });
   }
 };
+
+const reporteOfertasAgente = (req, res) =>{
+
+  
+
+}
 
 module.exports = { get, post, update, delte };
